Faster construction of items store.
ListStore sorting functionality sorts items when every one of them are
added making it very unperformant when adding a lot of items at once.
Creating new store and replacing it is not an option, but as sorting
direction/field cannot be changed we can utilize python's sorting
methods and sort list before adding it to a store.