Add a cache for Item objects
This reduces severity of leaks in GtkListStore by a very big amount.
Items added and then cleared from store currently are not dereferenced
correctly causing them to not be collected. Now after we cache object we
do not care about it that much anymore as we keep ONLY ONE copy of all
objects at any given time.
We still leak memory a bit, but now it is not in our control anymore.

Fixes #17